requirements.txt
# JARVIS Requirements
# Personal AI Desktop Agent for Akshay
# Complete build with all 10 additional features

# ==================== CORE ====================
groq>=0.4.0

# ==================== VOICE ====================
faster-whisper>=0.10.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
openWakeWord>=0.1.0

# ==================== PC AUTOMATION ====================
pyautogui>=0.9.54
pynput>=1.7.6
psutil>=5.9.6
pyperclip>=1.8.2

# ==================== BROWSER ====================
webbrowser (built-in)
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.40.0

# ==================== MEMORY ====================
chromadb>=0.4.22
sentence-transformers>=2.2.2

# ==================== MEDIA & OCR ====================
yt-dlp>=2023.12.30
Pillow>=10.1.0
pytesseract>=0.3.10

# ==================== PDF & DOCUMENTS ====================
PyMuPDF>=1.23.0

# ==================== SCHEDULING ====================
APScheduler>=3.10.4
schedule>=1.2.0

# ==================== WEB FRAMEWORK ====================
flask>=3.0.0
flask-socketio>=5.3.6
uvicorn>=0.25.0

# ==================== UTILS ====================
python-dotenv>=1.0.0
playsound>=1.3.0

# ==================== GIT ====================
GitPython>=3.1.40

# ==================== GPU MONITORING ====================
GPUtil>=1.4.0

# ==================== OPTIONAL ====================
# For full WhatsApp integration:
# whatsapp-web.js (requires Node.js)

# For Sarvam AI Kannada voice:
# requests (already included)

# Note: Tesseract OCR requires separate installation
# Download from: https://github.com/UB-Mannheim/tesseract/wiki
